refactor(admin): move backup progress types out of `create/`
Restore reports progress the same way and over largely the same vocabulary, so
these belong to the backup module rather than to creating one.
